LGitId>>hexString improvements
1. Pass a ByteArray as the return buffer instead of String in #hexString.
Using a String works in Squeak FFI as the result is guaranteed to be ascii, but it is bad practice in general.
In Threaded FFI Strings are utf8 encoded and then passed in a buffer that is discarded after the call completes, i.e. they can't be used to return values (which is good since the decoding hasn't been done).
2. Use git_oid_nfmt() instead of git_oid_fmt().  The latter is suspectable to buffer overruns, which will corrupt object memory.