LGitCommitBuilder>>commit_create:repo:update_ref:author:committer:message_encoding:message:tree:parent_count:parents: parent_count
parent_count is currently defined as an int, which is signed 32 bits.

As documented at https://libgit2.org/libgit2/#HEAD/group/commit/git_commit_create it is actually a size_t, which is unsigned 64 bits on a 64 bit platform.

In Squeak FFI this doesn't seem to matter, presumably it zero outs the top half when transferred as a 64 bit value, however Threaded FFI doesn't appear to do this.  The result is that git commits will occasionally crash the VM when using Threaded FFI.

Modify the function signature to use size_t.

Fixes: https://github.com/pharo-project/pharo/issues/5592